+++ src/model2/factories/index.ts
@@ -1,3 +1,4 @@
+export * from "./account_factory";
 export * from "./account_flags_factory";
 export * from "./account_thresholds_factory";
 export * from "./signer_factory";
